Refactor editing form out of /edit
