+++ app/routes/entries/$headword/MeaningEditingForm.tsx
@@ -2,9 +2,9 @@ import { Form } from "@remix-run/react"
 import { type MeaningType } from "~/components/Meaning"
 import { attributeEnum } from "~/components/editing/attributeEnum"
 import Button from "~/components/elements/Button"
-import { EditFormInput } from "./edit"
 import { useState } from "react"
 import { CanadianismTypeEnum } from "~/types/CanadianismTypeEnum"
+import { EditFormInput } from "./EntryEditingForm"
 
 interface MeaningEditingFormProps {
   headword: string
@@ -16,6 +16,7 @@ export default function MeaningEditingForm({
   meaning,
 }: MeaningEditingFormProps) {
   const {
+    definition,
     order,
     partofspeech,
     dagger,
@@ -23,6 +24,7 @@ export default function MeaningEditingForm({
     canadianism_type_comment,
   } = meaning
 
+  const [definitionValue, setDefinition] = useState(definition)
   const [orderValue, setOrder] = useState(order ?? "")
   const [daggerValue, setDagger] = useState(dagger)
   const [partOfSpeech, setPartOfSpeech] = useState(partofspeech)
@@ -33,10 +35,10 @@ export default function MeaningEditingForm({
   //   const [usageNoteValue, setUsageNote] = useState(usageNote)
 
   return (
-    <div>
+    <div className="my-16">
       <h3 className="text-4xl font-bold">Meaning: {meaning.order}</h3>
-      <Form method="post">
-        <div className="my-10 grid grid-cols-6">
+      <Form method="post" className="my-6">
+        <div className="grid grid-cols-7">
           <input
             type="hidden"
             name="attributeType"
@@ -44,6 +46,14 @@ export default function MeaningEditingForm({
           />
           <input type="hidden" name="id" value={meaning.id} />
           <input type="hidden" name="headword" value={headword} />
+          <EditFormInput
+            label="Definition: "
+            value={definitionValue}
+            name="definition"
+            onChangeFunction={setDefinition}
+            className="col-span-7"
+          />
+
           <EditFormInput
             label="Order: "
             name="order"
@@ -64,7 +74,7 @@ export default function MeaningEditingForm({
               name="canadianismType"
               value={canadianismType ?? ""}
               onChange={(e) => setCanadianismType(e.target.value)}
-              className="mx-2 p-1"
+              className="mx-2 border p-1"
             >
               <option value={""}>{"None"}</option>
               <option>{CanadianismTypeEnum.ONE_ORIGIN}</option>
@@ -75,7 +85,29 @@ export default function MeaningEditingForm({
               <option>{CanadianismTypeEnum.SIX_MEMORIAL}</option>
             </select>
           </label>
-          <Button type="submit" size="medium" className="col-start-3">
+          <label className="col-span-1">
+            Dagger:
+            <input
+              name="dagger"
+              checked={daggerValue}
+              type="checkbox"
+              onChange={(e) => setDagger(e.target.checked)}
+              className="mx-4"
+            />
+          </label>
+          <EditFormInput
+            label="Canadianism Type Comment:"
+            value={canadianismTypeComment || ""}
+            name="canadianismTypeComment"
+            type="textarea"
+            onChangeFunction={setCanadianismTypeComment}
+            className="col-span-7"
+          />
+          <Button
+            type="submit"
+            size="medium"
+            className="col-span-1 col-start-4"
+          >
             Update Meaning
           </Button>
         </div>
