+++ app/routes/entries/$headword/edit.tsx
@@ -1,15 +1,15 @@
 import { redirect, type ActionArgs, type LoaderArgs } from "@remix-run/node"
-import { Form, useCatch, useLoaderData } from "@remix-run/react"
+import { Form, Link, useCatch, useLoaderData } from "@remix-run/react"
 import invariant from "tiny-invariant"
 import { getEntryByHeadword, updateEntry } from "~/models/entry.server"
 import { getAttributeEnumFromFormInput } from "~/utils/generalUtils"
 import { attributeEnum } from "~/components/editing/attributeEnum"
 import { PageHeader } from "~/components/elements/PageHeader"
-import React, { useState } from "react"
 import Button from "~/components/elements/Button"
 import MeaningEditingForm from "./MeaningEditingForm"
 import { type LoadedDataType } from "."
-import { updateMeaning } from "~/models/update.server"
+import { addMeaningToEntry, updateMeaning } from "~/models/update.server"
+import EntryEditingForm from "./EntryEditingForm"
 
 export async function loader({ params }: LoaderArgs) {
   invariant(params.headword, "headword not found")
@@ -23,12 +23,14 @@ export async function loader({ params }: LoaderArgs) {
 
 export async function action({ request }: ActionArgs) {
   const data = Object.fromEntries(await request.formData())
-
   const type = getAttributeEnumFromFormInput(data.attributeType)
 
   switch (type) {
     case attributeEnum.ENTRY:
       await updateEntry(data)
+      break
+    case attributeEnum.ADD_MEANING:
+      await addMeaningToEntry(data)
       break
     case attributeEnum.MEANING:
       await updateMeaning(data)
@@ -37,161 +39,40 @@ export async function action({ request }: ActionArgs) {
       throw new Error("attribute enum unknown")
   }
 
-  return redirect(`/entries/${data.headword}`)
-}
-
-interface editHeadwordInputProps {
-  label: string
-  value: string
-  name: string
-  type?: "input" | "textarea" | "dropdown"
-  onChangeFunction: React.SetStateAction<any>
-  className?: string
-}
-
-export function EditFormInput(props: editHeadwordInputProps) {
-  let inputElm = getEditFormInput(props)
-  let { className, name, label } = props
-
-  return (
-    <div className={className}>
-      <label htmlFor={name} className="m-2">
-        {label}
-      </label>
-      {inputElm}
-    </div>
-  )
-}
-
-function getEditFormInput({
-  value,
-  name,
-  type,
-  onChangeFunction,
-}: editHeadwordInputProps) {
-  switch (type) {
-    case "textarea":
-      return (
-        <textarea
-          value={value}
-          name={name}
-          onChange={(e) => onChangeFunction(e.target.value)}
-          className="m-1 h-24 w-full border p-1"
-        ></textarea>
-      )
-    default:
-      return (
-        <input
-          value={value}
-          name={name}
-          type={type}
-          onChange={(e) => onChangeFunction(e.target.value)}
-          className="m-1 border p-1"
-        />
-      )
-  }
+  return redirect(`/entries/${data.headword}/edit`)
 }
 
 export default function EntryDetailsPage() {
   const data = useLoaderData<typeof loader>() as LoadedDataType
   const id = data.id
-  const [headword, setHeadword] = useState(data.headword)
-  const [spellingVariant, setSpellingVariant] = useState(
-    data.spelling_variants || ""
-  )
-  const [generalLabels, setGeneralLabels] = useState(data.general_labels || "")
-  const [handNote, setHandnote] = useState(data.fist_note || "")
-  const [dagger, setDagger] = useState(data.dagger)
-  const [etymology, setEtymology] = useState(data.etymology || "")
-  const [isLegacy, setIsLegacy] = useState(data.is_legacy)
-  const [isNonCanadian, setIsNonCanadian] = useState(data.no_cdn_conf)
+  const headword = data.headword
 
   return (
     <div className="w-3/4">
-      <PageHeader>Edit Headword: {data.headword}</PageHeader>
-      <Form method="post">
-        <div className="grid grid-cols-6">
-          <input type="hidden" name="id" value={id} />
+      <div className="flex flex-row items-center justify-between">
+        <PageHeader>Edit Headword: {headword}</PageHeader>
+        <Link
+          to={`/entries/${headword}`}
+          className="h-fit rounded border border-slate-700 bg-slate-600 p-2 text-white transition-colors duration-300 hover:bg-slate-500"
+        >
+          Return To Headword
+        </Link>
+        <Form
+          reloadDocument={true}
+          action={`/entries/${headword}/edit`}
+          method="post"
+        >
+          <Button>Add New Meaning</Button>
           <input
             type="hidden"
             name="attributeType"
-            value={attributeEnum.ENTRY}
+            value={attributeEnum.ADD_MEANING}
           />
-          <EditFormInput
-            label="headword: "
-            name="headword"
-            value={headword}
-            onChangeFunction={setHeadword}
-            className="col-span-2"
-          />
-          <EditFormInput
-            label="Spelling Variants: "
-            name="spellingVariant"
-            value={spellingVariant}
-            onChangeFunction={setSpellingVariant}
-            className="col-span-2"
-          />
-          <EditFormInput
-            label="General Labels: "
-            name="generalLabels"
-            value={generalLabels}
-            onChangeFunction={setGeneralLabels}
-            className="col-span-2"
-          />
-          <EditFormInput
-            label="Etymology: "
-            name="etymology"
-            value={etymology}
-            onChangeFunction={setEtymology}
-            className="col-span-2"
-          />
-          <label className="col-span-1 m-2">
-            Dagger
-            <input
-              name="dagger"
-              checked={dagger}
-              onChange={(e) => setDagger(e.target.checked)}
-              type="checkBox"
-              className="mx-3 border p-1"
-            />
-          </label>
-          <label className="col-span-1 m-2">
-            Legacy:
-            <input
-              name="isLegacy"
-              checked={isLegacy}
-              onChange={(e) => setIsLegacy(e.target.checked)}
-              type="checkBox"
-              className="mx-3 border p-1"
-            />
-          </label>
-          <label className="col-span-1 m-2">
-            Non Canadian:
-            <input
-              name="isNonCanadian"
-              checked={isNonCanadian}
-              onChange={(e) => setIsNonCanadian(e.target.checked)}
-              type="checkBox"
-              className="mx-3 border p-1"
-            />
-          </label>
-          <EditFormInput
-            label="Fist Note: "
-            name="fistNote"
-            value={handNote}
-            type="textarea"
-            onChangeFunction={setHandnote}
-            className="col-span-6 my-4"
-          />
-          <Button
-            className="col-span-2 col-start-3"
-            type="submit"
-            size="medium"
-          >
-            Submit
-          </Button>
-        </div>
-      </Form>
+          <input type="hidden" name="attributeID" value={id} />
+        </Form>
+      </div>
+      <EntryEditingForm data={data} />
+
       <div id="definitions">
         {data.meanings
           .sort((a, b) => (a.order || "").localeCompare(b.order || ""))
